Feature/Finalise automatic trainings channel (#66)
* Improve starts in time notation

* Add next training field

* Update all commands and put helper functions in seperate files

* Extend clear command so that it can also clear the trainings channel

* Lint
