+++ app/commands/admin/clear.js
@@ -9,8 +9,8 @@ module.exports = class ClearCommand extends Command {
             name: 'clearchannel',
             aliases: ['clear'],
             description: 'Clears given channel.',
-            details: 'Only channels #reports and #suggestions can be cleared. This will delete all messages but the ' +
-            'important information ones.',
+            details: 'Only channels #reports, #suggestions and #trainings can be cleared. This will delete all ' +
+                'messages but the important information ones.',
             examples: ['clear #suggestions'],
             clientPermissions: ['MANAGE_MESSAGES', 'ADD_REACTIONS', 'VIEW_CHANNEL', 'SEND_MESSAGES'],
             ownerOnly: true,
@@ -28,8 +28,11 @@ module.exports = class ClearCommand extends Command {
         const channels = guild.getData('channels')
         const suggestionsChannelId = channels.suggestionsChannel
         const reportsChannelId = channels.reportsChannel
-        if (channel.id !== suggestionsChannelId && channel.id !== reportsChannelId) {
-            return message.reply(`Can only clear <#${suggestionsChannelId}> or <#${reportsChannelId}>.`)
+        const trainingsChannelId = channels.trainingsChannel
+        if (channel.id !== suggestionsChannelId && channel.id !== reportsChannelId && channel.id !==
+            trainingsChannelId) {
+            return message.reply(`Can only clear <#${suggestionsChannelId}>, <#${reportsChannelId}> or <#${
+                trainingsChannelId}>.`)
         }
         const choice = await discordService.prompt(message.channel, message.author, await message.reply('Are you sure' +
             ` you would like to clear ${channel}?`))
@@ -37,8 +40,9 @@ module.exports = class ClearCommand extends Command {
             const guildMessages = guild.getData('messages')
             let messages
             do {
-                messages = await channel.messages.fetch({ after: channel.id === suggestionsChannelId ? guildMessages
-                    .firstSuggestionMessage : guildMessages.firstReportMessage })
+                const after = channel.id === suggestionsChannelId ? guildMessages.firstSuggestionMessage : channel
+                    .id === reportsChannelId ? guildMessages.firstReportMessage : guildMessages.trainingsMessage
+                messages = await channel.messages.fetch({ after })
                 if (messages.size > 0) {
                     try {
                         await channel.bulkDelete(messages)
