+++ app/jobs/announce-trainings.js
@@ -4,30 +4,33 @@ const userService = require('../services/user')
 const { MessageEmbed } = require('discord.js')
 const groupService = require('../services/group')
 const timeHelper = require('../helpers/time')
-const lodash = require('lodash')
+const pluralize = require('pluralize')
 
 const applicationConfig = require('../../config/application')
-
-const trainingTypes = {'Conductor': [], 'Customer Service Representative': []}
 
 module.exports = async guild  => {
     const channels = guild.getData('channels')
     const messages = guild.getData('messages')
     const channel = guild.guild.channels.cache.get(channels.trainingsChannel)
+
     const message = await channel.messages.fetch(messages.trainingsMessage)
     const trainings = (await applicationAdapter('get', `/v1/groups/${applicationConfig
-        .groupId}/trainings`)).data
-    trainings.sort((a, b) => a.date - b.date)
+        .groupId}/trainings?sort=date`)).data
     const authorIds = [...new Set(trainings.map(training => training.authorId))]
     const authors = await userService.getUsers(authorIds)
     const trainingsEmbed = getTrainingsEmbed(trainings, authors)
     message.edit(trainingsEmbed)
+
+    const now = new Date()
+    const nextTraining = trainings.find(training => new Date(training.date) > now)
+    const infoMessage = await channel.messages.fetch(messages.trainingInfoMessage)
+    infoMessage.embeds[0].fields[1].value = nextTraining ? getNextTrainingMessage(nextTraining, authors) : ':x: There' +
+        ' are currently no scheduled trainings.'
+    infoMessage.edit(infoMessage.embeds)
 }
 
 function getTrainingsEmbed (trainings, authors) {
-    const now = new Date()
-    const today = now.getDate()
-    const groupedTrainings = groupTrainingsByType(trainings)
+    const groupedTrainings = groupService.groupTrainingsByType(trainings)
     const types = Object.keys(groupedTrainings)
     const embed = new MessageEmbed()
         .setColor(applicationConfig.primaryColor)
@@ -40,18 +43,7 @@ function getTrainingsEmbed (trainings, authors) {
         if (typeTrainings.length > 0) {
             for (let j = 0; j < typeTrainings.length; j++) {
                 const training = typeTrainings[j]
-                const date = new Date(training.date)
-                const timeString = timeHelper.getTime(date)
-                const trainingDay = date.getDate()
-                const dateString = trainingDay === today ? 'Today' : trainingDay === today + 1 ? 'Tomorrow' : timeHelper
-                    .getDate(date)
-                const author = authors.find(author => author.id === training.authorId)
-                result += `**${dateString}** at **${timeString}** hosted by ${author.name}`
-                const hourDifference = date.getHours() - now.getHours()
-                if (trainingDay === today && hourDifference <= 5) {
-                    result += `\n> :alarm_clock: Starts in: **${hourDifference} hours**`
-                }
-                if (training.notes) result += `\n> ${training.notes}`
+                result += getTrainingMessage(training, authors)
                 if (j < typeTrainings.length) result += '\n'
             }
         } else {
@@ -62,12 +54,45 @@ function getTrainingsEmbed (trainings, authors) {
     return embed
 }
 
-function groupTrainingsByType (trainings) {
-    const result = {}
-    for (const training of trainings) {
-        const type = groupService.getRoleByAbbreviation(training.type)
-        if (!result[type]) result[type] = []
-        result[type].push(training)
+function getTrainingMessage (training, authors) {
+    const now = new Date()
+    const today = now.getDate()
+    const date = new Date(training.date)
+    const timeString = timeHelper.getTime(date)
+    const trainingDay = date.getDate()
+    const dateString = trainingDay === today ? 'Today' : trainingDay === today + 1 ? 'Tomorrow' : timeHelper
+        .getDate(date)
+    const author = authors.find(author => author.id === training.authorId)
+    let result = `:calendar_spiral: **${dateString}** at **${timeString}** hosted by ${author.name}`
+    const hourDifference = date.getHours() - now.getHours()
+    if (trainingDay === today && hourDifference <= 5) {
+        if (hourDifference === 0) {
+            const minuteDifference = date.getMinutes() - now.getMinutes()
+            if (minuteDifference >= 0) {
+                result += `\n> :alarm_clock: Starts in: **${minuteDifference} ${pluralize('minute',
+                    minuteDifference)}**`
+            } else {
+                result += `\n> :alarm_clock: Started **${-1 * minuteDifference} ${pluralize('minute',
+                    minuteDifference)}** ago`
+            }
+        } else {
+            result += `\n> :alarm_clock: Starts in: **${hourDifference} ${pluralize('hour', hourDifference)}**`
+        }
     }
-    return lodash.assign({}, trainingTypes, result)
+    if (training.notes) result += `\n> :notepad_spiral: ${training.notes}`
+    return result
+}
+
+function getNextTrainingMessage (training, authors) {
+    const now = new Date()
+    const today = now.getDate()
+    const date = new Date(training.date)
+    const timeString = timeHelper.getTime(date)
+    const trainingDay = date.getDate()
+    const dateString = trainingDay === today ? 'today' : trainingDay === today + 1 ? 'tomorrow' : timeHelper
+        .getDate(date)
+    const author = authors.find(author => author.id === training.authorId)
+    let result = `${training.type.toUpperCase()} **${dateString}** at **${timeString}** hosted by ${author.name}`
+    if (training.notes) result += `\n${training.notes}`
+    return result
 }
